individual ops
