+++ gsplat/strategy/ops.py
@@ -1,7 +1,12 @@
-from typing import Any, Callable, Dict, List, Union
+from typing import Any, Callable, DefaultDict, Dict, List, Tuple, Union
 
 import torch
+import torch.nn.functional as F
 from torch import Tensor
+
+from gsplat import quat_scale_to_covar_preci
+from gsplat.relocation import compute_relocation
+from gsplat.utils import normalized_quat_to_rotmat
 
 
 @torch.no_grad()
@@ -41,3 +46,274 @@ def _update_param_with_optimizer(
             optimizer.param_groups[i]["params"] = [p_new]
             optimizer.state[p_new] = p_state
             params[name] = p_new
+
+
+@torch.no_grad()
+def duplicate(
+    params: Union[Dict[str, torch.nn.Parameter], torch.nn.ParameterDict],
+    optimizers: Dict[str, torch.optim.Optimizer],
+    state: Dict[str, Tensor],
+    mask: Tensor,
+):
+    """Inplace duplicate the Gaussian with the given mask.
+
+    Args:
+        params: A dictionary of parameters.
+        optimizers: A dictionary of optimizers, each corresponding to a parameter.
+        mask: A boolean mask to duplicate the Gaussians.
+    """
+    device = mask.device
+    sel = torch.where(mask)[0]
+
+    def param_fn(name: str, p: Tensor) -> Tensor:
+        return torch.nn.Parameter(torch.cat([p, p[sel]]))
+
+    def optimizer_fn(key: str, v: Tensor) -> Tensor:
+        return torch.cat([v, torch.zeros((len(sel), *v.shape[1:]), device=device)])
+
+    # update the parameters and the state in the optimizers
+    _update_param_with_optimizer(param_fn, optimizer_fn, params, optimizers)
+    # update the extra running state
+    for k, v in state.items():
+        state[k] = torch.cat((v, v[sel]))
+
+
+@torch.no_grad()
+def split(
+    params: Union[Dict[str, torch.nn.Parameter], torch.nn.ParameterDict],
+    optimizers: Dict[str, torch.optim.Optimizer],
+    state: Dict[str, Tensor],
+    mask: Tensor,
+    revised_opacity: bool = False,
+):
+    """Inplace split the Gaussian with the given mask.
+
+    Args:
+        params: A dictionary of parameters.
+        optimizers: A dictionary of optimizers, each corresponding to a parameter.
+        mask: A boolean mask to split the Gaussians.
+        revised_opacity: Whether to use revised opacity formulation
+          from arXiv:2404.06109. Default: False.
+    """
+    device = mask.device
+    sel = torch.where(mask)[0]
+    rest = torch.where(~mask)[0]
+
+    scales = torch.exp(params["scales"])[sel]
+    quats = F.normalize(params["quats"], dim=-1)[sel]
+    opacities = torch.sigmoid(params["opacities"])[sel]
+    rotmats = normalized_quat_to_rotmat(quats)  # [N, 3, 3]
+    samples = torch.einsum(
+        "nij,nj,bnj->bni",
+        rotmats,
+        scales,
+        torch.randn(2, len(scales), 3, device=device),
+    )  # [2, N, 3]
+
+    def param_fn(name: str, p: Tensor) -> Tensor:
+        if name == "means3d":
+            p_split = (p[sel] + samples).reshape(-1, 3)  # [2N, 3]
+        elif name == "scales":
+            p_split = torch.log(scales / 1.6).repeat(2, 1)  # [2N, 3]
+        elif name == "opacities" and revised_opacity:
+            opacities = 1.0 - torch.sqrt(1.0 - opacities)
+            p_split = torch.logit(opacities).repeat(2)  # [2N]
+        else:
+            repeats = [2] + [1] * (p.dim() - 1)
+            p_split = p[sel].repeat(repeats)
+        p_new = torch.cat([p[rest], p_split])
+        p_new = torch.nn.Parameter(p_new)
+        return p_new
+
+    def optimizer_fn(key: str, v: Tensor) -> Tensor:
+        v_split = torch.zeros((2 * len(sel), *v.shape[1:]), device=device)
+        return torch.cat([v[rest], v_split])
+
+    # update the parameters and the state in the optimizers
+    _update_param_with_optimizer(param_fn, optimizer_fn, params, optimizers)
+    # update the extra running state
+    for k, v in state.items():
+        repeats = [2] + [1] * (v.dim() - 1)
+        v_new = v[sel].repeat(repeats)
+        state[k] = torch.cat((v[rest], v_new))
+
+
+@torch.no_grad()
+def remove(
+    params: Union[Dict[str, torch.nn.Parameter], torch.nn.ParameterDict],
+    optimizers: Dict[str, torch.optim.Optimizer],
+    state: Dict[str, Tensor],
+    mask: Tensor,
+):
+    """Inplace remove the Gaussian with the given mask.
+
+    Args:
+        params: A dictionary of parameters.
+        optimizers: A dictionary of optimizers, each corresponding to a parameter.
+        mask: A boolean mask to remove the Gaussians.
+    """
+    sel = torch.where(~mask)[0]
+
+    def param_fn(name: str, p: Tensor) -> Tensor:
+        return torch.nn.Parameter(p[sel])
+
+    def optimizer_fn(key: str, v: Tensor) -> Tensor:
+        return v[sel]
+
+    # update the parameters and the state in the optimizers
+    _update_param_with_optimizer(param_fn, optimizer_fn, params, optimizers)
+    # update the extra running state
+    for k, v in state.items():
+        state[k] = v[sel]
+
+
+@torch.no_grad()
+def reset_opa(
+    params: Union[Dict[str, torch.nn.Parameter], torch.nn.ParameterDict],
+    optimizers: Dict[str, torch.optim.Optimizer],
+    state: Dict[str, Tensor],
+    value: float,
+):
+    """Inplace reset the opacities to the given post-sigmoid value.
+
+    Args:
+        params: A dictionary of parameters.
+        optimizers: A dictionary of optimizers, each corresponding to a parameter.
+        value: The value to reset the opacities
+    """
+
+    def param_fn(name: str, p: Tensor) -> Tensor:
+        if name == "opacities":
+            opacities = torch.clamp(p, max=torch.logit(torch.tensor(value)).item())
+            return torch.nn.Parameter(opacities)
+        else:
+            raise ValueError(f"Unexpected parameter name: {name}")
+
+    def optimizer_fn(key: str, v: Tensor) -> Tensor:
+        return torch.zeros_like(v)
+
+    # update the parameters and the state in the optimizers
+    _update_param_with_optimizer(
+        param_fn, optimizer_fn, params, optimizers, names=["opacities"]
+    )
+
+
+@torch.no_grad()
+def relocate(
+    params: Union[Dict[str, torch.nn.Parameter], torch.nn.ParameterDict],
+    optimizers: Dict[str, torch.optim.Optimizer],
+    state: Dict[str, Tensor],
+    mask: Tensor,
+    min_opacity: float = 0.005,
+):
+    """Inplace relocate some dead Gaussians to the lives ones.
+
+    Args:
+        params: A dictionary of parameters.
+        optimizers: A dictionary of optimizers, each corresponding to a parameter.
+        mask: A boolean mask to indicates which Gaussians are dead.
+    """
+    opacities = torch.sigmoid(params["opacities"])
+
+    dead_indices = mask.nonzero(as_tuple=True)[0]
+    alive_indices = (~mask).nonzero(as_tuple=True)[0]
+    n = len(dead_indices)
+
+    # Sample for new GSs
+    eps = torch.finfo(torch.float32).eps
+    probs = opacities[alive_indices]
+    sampled_idxs = torch.multinomial(probs, n, replacement=True)
+    sampled_idxs = alive_indices[sampled_idxs]
+    new_opacities, new_scales = compute_relocation(
+        opacities=opacities[sampled_idxs],
+        scales=torch.exp(params["scales"])[sampled_idxs],
+        ratios=torch.bincount(sampled_idxs)[sampled_idxs] + 1,
+    )
+    new_opacities = torch.clamp(new_opacities, max=1.0 - eps, min=min_opacity)
+
+    def param_fn(name: str, p: Tensor) -> Tensor:
+        if name == "opacities":
+            p[sampled_idxs] = torch.logit(new_opacities)
+        elif name == "scales":
+            p[sampled_idxs] = torch.log(new_scales)
+        p[dead_indices] = p[sampled_idxs]
+        return torch.nn.Parameter(p)
+
+    def optimizer_fn(key: str, v: Tensor) -> Tensor:
+        v[sampled_idxs] = 0
+        return v
+
+    # update the parameters and the state in the optimizers
+    _update_param_with_optimizer(param_fn, optimizer_fn, params, optimizers)
+    # update the extra running state
+    for k, v in state.items():
+        v[sampled_idxs] = 0
+
+
+@torch.no_grad()
+def sample_add(
+    params: Union[Dict[str, torch.nn.Parameter], torch.nn.ParameterDict],
+    optimizers: Dict[str, torch.optim.Optimizer],
+    state: Dict[str, Tensor],
+    n: int,
+    min_opacity: float = 0.005,
+):
+    opacities = torch.sigmoid(params["opacities"])
+
+    eps = torch.finfo(torch.float32).eps
+    probs = opacities
+    sampled_idxs = torch.multinomial(probs, n, replacement=True)
+    new_opacities, new_scales = compute_relocation(
+        opacities=opacities[sampled_idxs],
+        scales=torch.exp(params["scales"])[sampled_idxs],
+        ratios=torch.bincount(sampled_idxs)[sampled_idxs] + 1,
+    )
+    new_opacities = torch.clamp(new_opacities, max=1.0 - eps, min=min_opacity)
+
+    def param_fn(name: str, p: Tensor) -> Tensor:
+        if name == "opacities":
+            p[sampled_idxs] = torch.logit(new_opacities)
+        elif name == "scales":
+            p[sampled_idxs] = torch.log(new_scales)
+        p = torch.cat([p, p[sampled_idxs]])
+        return torch.nn.Parameter(p)
+
+    def optimizer_fn(key: str, v: Tensor) -> Tensor:
+        v_new = torch.zeros((len(sampled_idxs), *v.shape[1:]), device=v.device)
+        return torch.cat([v, v_new])
+
+    # update the parameters and the state in the optimizers
+    _update_param_with_optimizer(param_fn, optimizer_fn, params, optimizers)
+    # update the extra running state
+    for k, v in state.items():
+        v_new = torch.zeros((len(sampled_idxs), *v.shape[1:]), device=v.device)
+        state[k] = torch.cat((v, v_new))
+
+
+@torch.no_grad()
+def inject_noise_to_position(
+    params: Union[Dict[str, torch.nn.Parameter], torch.nn.ParameterDict],
+    optimizers: Dict[str, torch.optim.Optimizer],
+    state: Dict[str, Tensor],
+    scaler: float,
+):
+    opacities = torch.sigmoid(params["opacities"])
+    scales = torch.exp(params["scales"])
+    covars, _ = quat_scale_to_covar_preci(
+        params["quats"],
+        scales,
+        compute_covar=True,
+        compute_preci=False,
+        triu=False,
+    )
+
+    def op_sigmoid(x, k=100, x0=0.995):
+        return 1 / (1 + torch.exp(-k * (x - x0)))
+
+    noise = (
+        torch.randn_like(params["means3d"])
+        * (op_sigmoid(1 - opacities)).unsqueeze(-1)
+        * scaler
+    )
+    noise = torch.bmm(covars, noise.unsqueeze(-1)).squeeze(-1)
+    params["means3d"].add_(noise)
